New approach to pseudo-modules
