+++ lib/browser.ts
@@ -1,8 +1,11 @@
-/**
- * @moduledefinition aws-crt
- *
+/*
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0.
+ */
+
+/**
+ * @packageDocumentation
+ * @module crt
  */
 
 // This is the entry point for the browser AWS CRT shim library
