+++ lib/browser/io.ts
@@ -1,6 +1,12 @@
-/**
+/*
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0.
+ */
+
+/**
+ *
+ * @packageDocumentation
+ * @module IO
  */
 
 export { TlsVersion, SocketType, SocketDomain } from "../common/io";
@@ -9,7 +15,6 @@ import { SocketType, SocketDomain } from "../common/io";
 /**
  * @return false, as ALPN is not configurable from the browser
  *
- * @module aws-crt
  * @category TLS
 */
 export function is_alpn_available(): boolean {
@@ -22,8 +27,7 @@ type BodyData = string | object | ArrayBuffer | ArrayBufferView | Blob | File;
  * Wrapper for any sort of body data in requests. As the browser does not implement streaming,
  * this is merely an interface wrapper around a memory buffer.
  *
- * @module aws-crt
- * @category I/O
+ * @category IO
  */
 export class InputStream {
     constructor(public data: BodyData) {
@@ -35,8 +39,7 @@ export class InputStream {
  * Represents resources required to bootstrap a client connection, provided as
  * a stub for the browser API
  *
- * @module aws-crt
- * @category I/O
+ * @category IO
  */
 export class ClientBootstrap { };
 
@@ -44,7 +47,6 @@ export class ClientBootstrap { };
  * Options for creating a {@link ClientTlsContext}. Provided as a stub for
  * browser API.
  *
- * @module aws-crt
  * @category TLS
  */
 export type TlsContextOptions = any;
@@ -53,7 +55,6 @@ export type TlsContextOptions = any;
  * TLS options that are unique to a given connection using a shared TlsContext.
  * Provided as a stub for browser API.
  *
- * @module aws-crt
  * @category TLS
  */
 export class TlsConnectionOptions {
@@ -66,7 +67,6 @@ export class TlsConnectionOptions {
  * TLS context used for TLS communications over sockets. Provided as a
  * stub for the browser API
  *
- * @module aws-crt
  * @category TLS
  */
 export abstract class TlsContext {
@@ -77,7 +77,6 @@ export abstract class TlsContext {
  * TLS context used for client TLS communications over sockets. Provided as a
  * stub for the browser API
  *
- * @module aws-crt
  * @category TLS
  */
 export class ClientTlsContext extends TlsContext {
@@ -90,8 +89,7 @@ export class ClientTlsContext extends TlsContext {
  * Standard Berkeley socket style options.
  *
  * Provided for compatibility with nodejs, but this version is largely unused.
- * @module aws-crt
- * @category I/O
+ * @category Network
 */
 export class SocketOptions {
     constructor(
