+++ lib/common/resource_safety.ts
@@ -1,13 +1,17 @@
-/**
+/*
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0.
+ */
+
+/**
+ * @packageDocumentation
+ * @module crt
  */
 
 /**
  * If you have a resource that you want typescript to enforce close is implemented
  * and/or you want to use the below 'using' function, then implement this interface.
  *
- * @module aws-crt
  * @category System
  */
 export interface ResourceSafe {
@@ -25,7 +29,6 @@ export interface ResourceSafe {
  * });
  * ```
  *
- * @module aws-crt
  * @category System
  */
 export async function using<T extends ResourceSafe>(resource: T, func: (resource: T) => void) {
