+++ lib/native/auth.ts
@@ -1,6 +1,14 @@
-/**
+/*
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0.
+ */
+
+/**
+ * Module for AWS Authentication logic - signing http requests, events, chunks, etc...
+ *
+ * @packageDocumentation
+ * @module auth
+ * @preferred
  */
 
 import crt_native from './binding';
@@ -16,7 +24,6 @@ type StringLike = string | ArrayBuffer | DataView;
 /**
  * AWS signing algorithm enumeration.
  *
- * @module aws-crt
  * @category Auth
  */
 export enum AwsSigningAlgorithm {
@@ -30,7 +37,6 @@ export enum AwsSigningAlgorithm {
 /**
  * AWS signature type enumeration.
  *
- * @module aws-crt
  * @category Auth
  */
 export enum AwsSignatureType {
@@ -53,7 +59,6 @@ export enum AwsSignatureType {
  * Some services use special values (e.g. 'UNSIGNED-PAYLOAD') when the body
  * is not being signed in the usual way.
  *
- * @module aws-crt
  * @category Auth
  */
 export enum AwsSignedBodyValue {
@@ -73,7 +78,6 @@ export enum AwsSignedBodyValue {
 /**
  * AWS signed body header enumeration.
  *
- * @module aws-crt
  * @category Auth
  */
 export enum AwsSignedBodyHeaderType {
@@ -89,7 +93,6 @@ export enum AwsSignedBodyHeaderType {
  *
  * We don't currently expose an interface for fetching credentials from Javascript.
  *
- * @module aws-crt
  * @category Auth
  */
 /* Subclass for the purpose of exposing a non-NativeHandle based API */
@@ -119,10 +122,77 @@ export class AwsCredentialsProvider extends crt_native.AwsCredentialsProvider {
  * AwsSigningConfig is immutable.
  * It is good practice to use a new config for each signature, or the date might get too old.
  *
- * @module aws-crt
  * @category Auth
  */
-export type AwsSigningConfig = crt_native.AwsSigningConfig;
+export interface AwsSigningConfig {
+    /** Which signing process to invoke */
+    algorithm: AwsSigningAlgorithm;
+
+    /** What kind of signature to compute */
+    signature_type: AwsSignatureType;
+
+    /** Credentials provider to fetch signing credentials with */
+    provider: AwsCredentialsProvider;
+
+    /** The region to sign against */
+    region: string;
+
+    /** Name of service to sign a request for */
+    service?: string;
+
+    /**
+     * Date and time to use during the signing process. If not provided then
+     * the current time in UTC is used. Naive dates (lacking timezone info)
+     * are assumed to be in local time
+     */
+    date?: Date;
+
+    /**
+     * Headers to skip when signing.
+     *
+     * Skipping auth-required headers will result in an unusable signature.
+     * Headers injected by the signing process are not skippable.
+     * This function does not override the internal check function
+     * (x-amzn-trace-id, user-agent), but rather supplements it.
+     * In particular, a header will get signed if and only if it returns
+     * true to both the internal check (skips x-amzn-trace-id, user-agent)
+     * and is found in this list (if defined)
+     */
+    header_blacklist?: string[];
+
+    /**
+     * Set true to double-encode the resource path when constructing the
+     * canonical request. By default, all services except S3 use double encoding.
+     */
+    use_double_uri_encode?: boolean;
+
+    /**
+     * Whether the resource paths are normalized when building the canonical request.
+     */
+    should_normalize_uri_path?: boolean;
+
+    /**
+     * Should the session token be omitted from the signing process?  This should only be
+     * true when making a websocket handshake with IoT Core.
+     */
+    omit_session_token?: boolean;
+
+    /**
+     * Value to use as the canonical request's body value.
+     *
+     * Typically, this is the SHA-256 of the payload, written as lowercase hex.
+     * If this has been precalculated, it can be set here.
+     * Special values used by certain services can also be set (see {@link AwsSignedBodyValue}).
+     * If undefined (the default), the typical value will be calculated from the payload during signing.
+     */
+    signed_body_value?: string;
+
+    /** Controls what header, if any, should be added to the request, containing the body value */
+    signed_body_header?: AwsSignedBodyHeaderType;
+
+    /** Query param signing only: how long the pre-signed URL is valid for */
+    expiration_in_seconds?: number;
+}
 
 /**
  * Perform AWS HTTP request signing.
@@ -151,7 +221,6 @@ export type AwsSigningConfig = crt_native.AwsSigningConfig;
  *       {@link HttpRequest}. The future will contain an exception
  *       if the signing process fails.
  *
- * @module aws-crt
  * @category Auth
  */
 export async function aws_sign_request(request: HttpRequest, config: AwsSigningConfig): Promise<HttpRequest> {
