+++ lib/native/checksums.ts
@@ -1,6 +1,11 @@
-/**
+/*
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0.
+ */
+
+/**
+ * @packageDocumentation
+ * @module crypto
  */
 
  import crt_native from './binding';
@@ -13,7 +18,6 @@
  * @param data The data to checksum
  * @param previous previous crc32 checksum result. Used if you are buffering large input.
  *
- * @module aws-crt
  * @category Crypto
  */
 export function crc32(data: Hashable, previous?: number): number {
@@ -26,7 +30,6 @@ export function crc32(data: Hashable, previous?: number): number {
  * @param data The data to checksum
  * @param previous previous crc32c checksum result. Used if you are buffering large input.
  *
- * @module aws-crt
  * @category Crypto
  */
  export function crc32c(data: Hashable, previous?: number): number {
